+++ test/dart-python--datastore-compatibility/dart/bin/main.dart
@@ -16,114 +16,34 @@ import 'package:gcloud/db.dart' as db;
 
 
 class AllDataTypesModelMixin {
+  @BoolProperty()
   bool boolProp;
+
+  @IntProperty()
   int intProp;
+
+  @StringProperty()
   String stringProp;
+
+  @ModelKeyProperty()
   Key keyProp;
+
+  @BlobProperty()
   List<int> blobProp;
+
+  @DateTimeProperty()
   DateTime dateProp;
 
+  @StringListProperty()
   List<String> stringListProp;
 }
 
-@ModelMetadata(const NormalModelDesc())
+@Kind()
 class NormalModel extends Model with AllDataTypesModelMixin { }
 
-@ModelMetadata(const PolyModelDesc())
-class PolyModel extends db.PolyModel with AllDataTypesModelMixin { }
-
-@ModelMetadata(const ExpandoModelDesc())
+@Kind()
 class ExpandoModel extends db.ExpandoModel with AllDataTypesModelMixin { }
 
-@ModelMetadata(const SubPolyModelDesc1())
-class SubPolyModel1 extends PolyModel {
-  String additionalSPM1;
-}
-
-@ModelMetadata(const SubPolyModelDesc2())
-class SubPolyModel2 extends PolyModel {
-  String additionalSPM2;
-}
-
-class AllDataTypesModelDescriptionMixin {
-  // These need to be copy & pasted to classes down, because const constructors
-  // on mixins are not supported yet.
-
-  final boolProp = const BoolProperty();
-  final intProp = const IntProperty();
-  final stringProp = const StringProperty();
-  final keyProp = const ModelKeyProperty();
-  final blobProp = const BlobProperty();
-  final dateProp = const DateTimeProperty();
-
-  final stringListProp = const StringListProperty();
-}
-
-class NormalModelDesc extends ModelDescription {
-  final id = const IntProperty();
-
-  // Copy&Pasted from [AllDataTypesModelDescriptionMixin]
-  // [mixin not supported for const classes]
-  final boolProp = const BoolProperty();
-  final intProp = const IntProperty();
-  final stringProp = const StringProperty();
-  final keyProp = const ModelKeyProperty();
-  final blobProp = const BlobProperty();
-  final dateProp = const DateTimeProperty();
-  final stringListProp = const StringListProperty();
-
-  const NormalModelDesc() : super('NormalModel');
-}
-
-class PolyModelDesc extends db.PolyModelDescription {
-  static String PolyModelName = 'PolyModel';
-
-  final id = const IntProperty();
-
-  // Copy&Pasted from [AllDataTypesModelDescriptionMixin]
-  // [mixin not supported for const classes]
-  final boolProp = const BoolProperty();
-  final intProp = const IntProperty();
-  final stringProp = const StringProperty();
-  final keyProp = const ModelKeyProperty();
-  final blobProp = const BlobProperty();
-  final dateProp = const DateTimeProperty();
-  final stringListProp = const StringListProperty();
-
-  const PolyModelDesc();
-}
-
-class ExpandoModelDesc extends db.ExpandoModelDescription {
-  final id = const IntProperty();
-
-  // Copy&Pasted from [AllDataTypesModelDescriptionMixin]
-  // [mixin not supported for const classes]
-  final boolProp = const BoolProperty();
-  final intProp = const IntProperty();
-  final stringProp = const StringProperty();
-  final keyProp = const ModelKeyProperty();
-  final blobProp = const BlobProperty();
-  final dateProp = const DateTimeProperty();
-  final stringListProp = const StringListProperty();
-
-  const ExpandoModelDesc() : super('ExpandoModel');
-}
-
-class SubPolyModelDesc1 extends PolyModelDesc {
-  static String PolyModelName = 'SubPolyModel1';
-
-  final additionalSPM1 = const StringProperty();
-
-  const SubPolyModelDesc1() : super();
-}
-
-class SubPolyModelDesc2 extends PolyModelDesc {
-  static String PolyModelName = 'SubPolyModel2';
-
-  final additionalSPM2 = const StringProperty();
-
-  const SubPolyModelDesc2() : super();
-}
 
 final DateTime EndOfYear = new DateTime.utc(2014, 12, 31);
 customDate(int i) => EndOfYear.add(new Duration(hours: i));
@@ -151,40 +71,22 @@ void verifyData(DatastoreDB db, AllDataTypesModelMixin model, int i) {
 
 runTests(bool writingMode, DatastoreDB db) {
   var key = db.emptyKey.append(NormalModel, id: 99);
-  var pkey = key.append(PolyModel, id: 99);
-  var subpkey1 = key.append(PolyModel, id: 100);
-  var subpkey2 = key.append(PolyModel, id: 101);
   var ekey = key.append(ExpandoModel, id: 102);
 
   if (writingMode) {
     test('writing-test', () {
       var normalObj = new NormalModel();
-      var pm = new PolyModel();
-      var spm1 = new SubPolyModel1();
-      var spm2 = new SubPolyModel2();
       var em = new ExpandoModel();
 
-      pm.parentKey = key;
-      spm1.parentKey = key;
-      spm2.parentKey = key;
       em.parentKey = key;
 
       normalObj.id = 99;
-      pm.id = 99;
-      spm1.id = 100;
-      spm2.id = 101;
       em.id = 102;
 
-      spm1.additionalSPM1 = 'abc';
-      spm2.additionalSPM2 = 'xyz';
-
       fillData(db, normalObj, 1);
-      fillData(db, pm, 2);
-      fillData(db, spm1, 3);
-      fillData(db, spm2, 4);
       fillData(db, em, 5);
 
-      db.commit(inserts: [normalObj, pm, spm1, spm2, em]).then(expectAsync((_) {
+      db.commit(inserts: [normalObj, em]).then(expectAsync((_) {
         print('done');
       }));
     });
@@ -192,42 +94,24 @@ runTests(bool writingMode, DatastoreDB db) {
 
   if (!writingMode) {
     test('reading-test', () {
-      db.lookup([key, pkey, subpkey1, subpkey2, ekey])
+      db.lookup([key, ekey])
           .then(expectAsync((List<Model> models) {
-        expect(models, hasLength(5));
+        expect(models, hasLength(2));
 
         NormalModel model = models[0];
-        PolyModel pmodel = models[1];
-        SubPolyModel1 subpmodel1 = models[2];
-        SubPolyModel2  subpmodel2 = models[3];
-        ExpandoModel emodel = models[4];
+        ExpandoModel emodel = models[1];
 
         expect(model, isNotNull);
-        expect(pmodel, isNotNull);
-        expect(subpmodel1, isNotNull);
-        expect(subpmodel2, isNotNull);
         expect(emodel, isNotNull);
 
         expect(model.id, equals(99));
-        expect(pmodel.id, equals(99));
-        expect(subpmodel1.id, equals(100));
-        expect(subpmodel2.id, equals(101));
         expect(emodel.id, equals(102));
 
         expect(model.parentKey, equals(db.emptyKey));
-        expect(pmodel.parentKey, equals(key));
-        expect(subpmodel1.parentKey, equals(key));
-        expect(subpmodel2.parentKey, equals(key));
         expect(emodel.parentKey, equals(key));
 
         verifyData(db, model, 1);
-        verifyData(db, pmodel, 2);
-        verifyData(db, subpmodel1, 3);
-        verifyData(db, subpmodel2, 4);
         verifyData(db, emodel, 5);
-
-        expect(subpmodel1.additionalSPM1, equals('abc'));
-        expect(subpmodel2.additionalSPM2, equals('xyz'));
       }));
     });
   }