+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.2.1 2014-11-04
+
+* Small bugfix in storage API implementation
+* respect DART_PUB_SERVE url only in developer mode
+* sync db/datastore tests to the ones used in gcloud
+
 ## 0.2.0 2014-11-03
 
 * Use datastore/db APIs from package:gcloud